Added interface to model for easier interaction.
Something that got mentioned in class is that we shouldn't create a MusicPiece directly - there should be a level of abstraction between MusicPieceInterface and making a MusicPiece.

Another thing mentioned in class is that we shouldn't be playing "notes" per se. We should be playing 'playabale objects' - might include notes, but will also include decorated notes, etc.